backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1